Move ItemRepository.kt to repository module
