+++ app/src/main/java/com/example/firstapp/viewmodel/ItemViewModel.kt
@@ -5,7 +5,7 @@ import com.example.firstapp.groupie.GroupieItem
 import com.example.firstapp.model.Item
 import com.example.firstapp.model.ItemGold
 import com.example.firstapp.model.ItemImage
-import com.example.firstapp.repository.ItemRepository
+import com.example.firstapp.data.repository.ItemRepository
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import dagger.hilt.android.lifecycle.HiltViewModel
