+++ data/repository/src/main/java/com/example/firstapp/data/repository/ItemRepository.kt
@@ -1,11 +1,11 @@
-package com.example.firstapp.repository
+package com.example.firstapp.data.repository
 
 import com.example.firstapp.data.db.dao.ItemDao
 import com.example.firstapp.model.Item
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 
-class ItemRepository(var itemDao: com.example.firstapp.data.db.dao.ItemDao) {
+class ItemRepository(var itemDao: ItemDao) {
     suspend fun getItems(): List<Item> = withContext(Dispatchers.IO) {
         return@withContext itemDao.selectAll()
     }
